refactor(ccusage): split codex parser worker
Move Codex JSONL parsing and worker fan-out into adapter/codex/parser.ts so the public adapter index only handles report aggregation and pricing wiring.

The focused Codex adapter tests and package typecheck confirm the extracted parser keeps the same behavior.
